requirements.txt
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.3.5
pre-commit==2.5.1
pytorch-lightning==0.7.3
scipy==1.4.1
scikit-learn==0.22.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
albumentations==0.4.5
pretrainedmodels==0.7.4


eventlet==0.24.1
flask==1.0.2
flask-cors==3.0.7
flask-login==0.4.1
flask-restplus==0.12.1
flask-mongoengine==0.9.5
werkzeug==0.16.1
cython
scikit-image
requests
google_images_download==2.5.0
watchdog==0.8.3
pytest==3.9.3
pytest-ordering==0.6
imantics==0.1.9
flask-socketio==3.3.2
celery==4.2.2
Shapely==1.7.0
Pillow
matplotlib
keras==2.1.1
h5py
imgaug
IPython[all]
jupyter
redis
